ASH-57: Hide the expand chevron on scholar requests with no extra details
Only show the control when a request has form answers or attachments, and render those extras with the shared human-readable labels.
