+++ apps/scholar/__mocks__/lucide-react.js
@@ -20,8 +20,10 @@ const icons = {
   Calendar: createMockIcon('Calendar'),
   CheckCircle: createMockIcon('CheckCircle'),
   CheckSquare: createMockIcon('CheckSquare'),
+  ChevronDown: createMockIcon('ChevronDown'),
   ChevronLeft: createMockIcon('ChevronLeft'),
   ChevronRight: createMockIcon('ChevronRight'),
+  ChevronUp: createMockIcon('ChevronUp'),
   Download: createMockIcon('Download'),
   Eye: createMockIcon('Eye'),
   EyeOff: createMockIcon('EyeOff'),
